Safari-611.1.10/LayoutTests/fast/dom/Window/watchos/window-get-real-screen-dimensions.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../resources/js-test.js"></script>
</head>
<body></body>
<script>
shouldBe("screen.width", "320");
shouldBe("screen.height", "568");
</script>
</html>
